+++ chooloolib/src/main/AndroidManifest.xml
@@ -4,9 +4,14 @@
 
     <application>
         <activity android:name=".ui.permissions.PermissionsActivity" />
-        
+        <activity
+            android:name=".ui.call.CallActivity"
+            android:screenOrientation="fullSensor" />
+
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
     </application>
+
+
 </manifest>
