+++ chooloolib/src/main/java/com/chooloo/www/chooloolib/ui/call/CallContract.kt
@@ -1,12 +1,11 @@
-package com.chooloo.www.koler.ui.call
+package com.chooloo.www.chooloolib.ui.call
 
 import android.net.Uri
 import android.view.KeyEvent
-import com.chooloo.www.chooloolib.data.call.Call
 import com.chooloo.www.chooloolib.interactor.callaudio.CallAudioInteractor
 import com.chooloo.www.chooloolib.interactor.calls.CallsInteractor
 import com.chooloo.www.chooloolib.ui.base.BaseContract
-import com.chooloo.www.koler.ui.widgets.CallActions
+import com.chooloo.www.chooloolib.ui.widgets.CallActions
 
 interface CallContract : BaseContract {
     interface View : BaseContract.View {
