+++ chooloolib/src/main/java/com/chooloo/www/chooloolib/ui/callitems/CallItemsController.kt
@@ -1,7 +1,7 @@
-package com.chooloo.www.koler.ui.callitems
+package com.chooloo.www.chooloolib.ui.callitems
 
+import com.chooloo.www.chooloolib.adapter.CallItemsAdapter
 import com.chooloo.www.chooloolib.data.call.Call
-import com.chooloo.www.koler.adapter.CallItemsAdapter
 import com.chooloo.www.chooloolib.ui.list.ListController
 
 class CallItemsController<V : CallItemsContract.View>(view: V) :
