+++ chooloolib/src/main/java/com/chooloo/www/chooloolib/ui/contacts/ContactsController.kt
@@ -3,7 +3,6 @@ package com.chooloo.www.chooloolib.ui.contacts
 import com.chooloo.www.chooloolib.R
 import com.chooloo.www.chooloolib.adapter.ContactsAdapter
 import com.chooloo.www.chooloolib.data.account.ContactAccount
-import com.chooloo.www.chooloolib.ui.contact.ContactFragment
 import com.chooloo.www.chooloolib.ui.list.ListController
 
 open class ContactsController<V : ContactsContract.View>(view: V) :
