+++ chooloolib/src/main/java/com/chooloo/www/chooloolib/ui/contacts/ContactsFragment.kt
@@ -2,8 +2,6 @@ package com.chooloo.www.chooloolib.ui.contacts
 
 import com.chooloo.www.chooloolib.adapter.ContactsAdapter
 import com.chooloo.www.chooloolib.data.account.ContactAccount
-import com.chooloo.www.chooloolib.ui.base.BottomFragment
-import com.chooloo.www.chooloolib.ui.contact.ContactFragment
 import com.chooloo.www.chooloolib.ui.list.ListFragment
 
 open class ContactsFragment :
