+++ chooloolib/src/main/res/layout/call.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layoutDescription="@xml/call_scene"
-    tools:context=".ui.call.CallActivity">
+    tools:context="com.chooloo.www.chooloolib.ui.call.CallActivity">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/call_middle_guideline"
@@ -82,7 +82,7 @@
         app:layout_constraintTop_toBottomOf="@id/call_name_text" />
 
     <!-- call actions -->
-    <com.chooloo.www.koler.ui.widgets.CallActions
+    <com.chooloo.www.chooloolib.ui.widgets.CallActions
         android:id="@+id/call_actions"
         android:layout_width="match_parent"
         android:layout_height="wrap_content" />
