+++ chooloolib/src/main/res/xml-land/call_scene.xml
