+++ koler/src/main/AndroidManifest.xml
@@ -61,20 +61,13 @@
             </intent-filter>
         </activity>
 
-        <activity
-            android:name="com.chooloo.www.koler.ui.call.CallActivity"
-            android:screenOrientation="fullSensor" />
-
-
+        
         <service
             android:name="com.chooloo.www.chooloolib.service.CallService"
             android:exported="false"
             android:foregroundServiceType="phoneCall"
             android:permission="android.permission.BIND_INCALL_SERVICE">
 
-            <meta-data
-                android:name="call_activity"
-                android:value="com.chooloo.www.koler.ui.call.CallActivity" />
             <meta-data
                 android:name="android.telecom.IN_CALL_SERVICE_UI"
                 android:value="true" />
