--- koler/src/main/res/xml/tab_state_animator.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_selected="true">
-        <set>
-            <objectAnimator
-                android:duration="0"
-                android:propertyName="color"
-                android:valueTo="?android:textColor" />
-        </set>
-    </item>
-    <item>
-        <set>
-            <objectAnimator
-                android:duration="0"
-                android:propertyName="color"
-                android:valueTo="?colorSecondaryVariant" />
-        </set>
-    </item>
-</selector>
